+++ lib/avro_turf/confluent_schema_registry.rb
@@ -35,7 +35,8 @@ class AvroTurf::ConfluentSchemaRegistry
       client_key_pass: client_key_pass,
       client_cert_data: client_cert_data,
       client_key_data: client_key_data,
-      connect_timeout: connect_timeout
+      connect_timeout: connect_timeout,
+      dns_timeouts: connect_timeout
     )
   end