+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Set `idempotent: true` for the request except one that registers a new schema (#199)
+- Use `connect_timeout` for `Excon`'s `dns_timeouts` that set the timeout for the connection to the Domain Name Server (#201)
 
 ## v1.12.0